+++ FunnyMoneyClicker/Assets/Scripts/UpgradeData.cs
@@ -6,7 +6,7 @@ public class UpgradeData
 {
     [Header("Upgrade Settings")]
     public string upgradeName; // The display name of the upgrade (e.g. "Auto Clicker", "Miner", "Factory")
-    public float baseCost = 10f; // Starting price for the first level of this upgrade
+    public double baseCost = 10f; // Starting price for the first level of this upgrade
     public float baseProduction = 1f; // How much money (or resources) this upgrade produces per cycle at level 1
 
     [HideInInspector] public float costIncreaseRate; // Multiplier for cost growth each level (e.g. cost *= 1.15)
